requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pandas==2.2.2
datasets==2.21.0
google-generativeai==0.8.4
pydantic==2.8.2
pytest==8.3.2
httpx==0.27.0
